+++ utilities/image_operations.py
